+++ tests/tests-core/src/main/java/com/gentics/mesh/core/schema/field/DateFieldMigrationTest.java
@@ -219,7 +219,7 @@ public class DateFieldMigrationTest extends AbstractFieldMigrationTest implement
 	public void testChangeToNumber() throws Exception {
 		changeType(CREATEDATE, FILL, FETCH, CREATENUMBER, (container, name) -> {
 			assertThat(container.getNumber(name)).as(NEWFIELD).isNotNull();
-			assertThat(container.getNumber(name).getNumber()).as(NEWFIELDVALUE).isEqualTo(DATEVALUE);
+			assertThat(container.getNumber(name).getNumber().longValue()).as(NEWFIELDVALUE).isEqualTo(DATEVALUE);
 		});
 	}
